requirements.txt
    Flask==2.3.2
    Pillow>=10.3.0
    opencv-python==4.9.0.80
    numpy==1.26.4
    boto3==1.34.80
    python-dotenv==1.0.1
    pytesseract==0.3.10
    
